safari-600.7-branch/LayoutTests/java/lc3/JavaObject/JavaObjectToChar-002-expected.txt

LiveConnect 3.0 JavaScript to Java Data Type Conversion JavaObject to char

PASS dt.PUB_DOUBLE_REPRESENTATION = -0.6;dt.setChar( dt ); dt.PUB_CHAR is 0
PASS dt.PUB_DOUBLE_REPRESENTATION = -0.6;dt.setChar( dt ); dt.getChar() is 0
PASS typeof dt.getChar() is 'number'
PASS successfullyParsed is true

TEST COMPLETE
 
